backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.19
asyncpg==0.27.0
pydantic==1.10.11
passlib[bcrypt]==1.7.4
PyJWT==2.8.0